refactor(cardwired): split gpu helpers into multiple files
