+++ crates/cardwire-daemon/src/core/gpu/mod.rs
@@ -1,9 +1,12 @@
+mod default_gpu;
+mod device_info;
+mod display;
 mod egl;
 mod enumerator;
-mod helpers;
 mod models;
 mod vulkan;
 
+pub use default_gpu::check_default_drm_class;
+pub use display::external_display_connected;
 pub use enumerator::GpuEnumerator;
-pub use helpers::{check_default_drm_class, external_display_connected};
 pub use models::{DbusGpuDevice, GpuDevice, GpuVendor, PowerState};
