+++ crates/cardwire-daemon/src/core/mod.rs
@@ -1,5 +1,6 @@
 pub mod errors;
 pub mod gpu;
 pub mod pci;
+pub mod procfs;
 
 pub mod inode;
